minimize header
fixes build on i686
